🪲fix<api deletar produto>: - Correção de bug para excluir produto, feito a verificação se o objeto imgs é um array.

+++ src/modulos/produto/produto.service.ts
@@ -109,7 +109,7 @@ export class ProdutoService {
     async deletarProduto(data: { produtoId: number }) {
         const produto = await this.produtoRepository.findOne({ where: { id: data.produtoId } });
         if (!produto) throw new NotFoundException('Produto não encontrado.')
-        await Promise.all(produto.imgs.map(async (img) => { await this.s3Service.delete(img.Key, img.Bucket); }));
+        if(Array.isArray(produto.imgs)) await Promise.all(produto.imgs.map(async (img) => { await this.s3Service.delete(img.Key, img.Bucket); }));
         return this.produtoRepository.delete(data.produtoId);
     }
